separate timeline from article status again
