+++ src/types/article-status.ts
@@ -1,31 +1,4 @@
-
-type TimelineEventBasic = {
-  name: string,
-  date: string,
-};
-
-type TimelineEventSimple = TimelineEventBasic & {
-  link?: undefined,
-  eventDescription?: undefined,
-};
-
-type TimelineEventWithLink = TimelineEventBasic & {
-  link: {
-    text: string,
-    url?: string,
-  },
-  eventDescription?: undefined,
-};
-
-type TimelineEventWithDescription = TimelineEventBasic & {
-  link?: undefined,
-  eventDescription: string,
-};
-
-export type TimelineEvent = TimelineEventSimple | TimelineEventWithLink | TimelineEventWithDescription;
-
 export type ArticleStatus = {
   type: 'Reviewed Preprint' | 'Revised Preprint',
   isPreview: boolean,
-  timeline: TimelineEvent[],
 };
