+++ src/types/article-timeline.ts
@@ -1,0 +1,24 @@
+type TimelineEventBasic = {
+  name: string,
+  date: string,
+};
+
+type TimelineEventSimple = TimelineEventBasic & {
+  link?: undefined,
+  eventDescription?: undefined,
+};
+
+type TimelineEventWithLink = TimelineEventBasic & {
+  link: {
+    text: string,
+    url?: string,
+  },
+  eventDescription?: undefined,
+};
+
+type TimelineEventWithDescription = TimelineEventBasic & {
+  link?: undefined,
+  eventDescription: string,
+};
+
+export type TimelineEvent = TimelineEventSimple | TimelineEventWithLink | TimelineEventWithDescription;
